Add support for full networks and named applications
The existing rules code doesn't allow for a couple of handy features
in UFW, being able to drop netblocks and using pre canned configuration.

Although this system results in largely duplicated code it was better
then the alternative I tried which was to mangle the existing ufw_rules
code to handle all three.

If the new variables are empty, they are skipped. If they contain a list
(optionally with a rule: deny) they are run (the two configurations I've
tested).